Denis Lakic, Deni - contribution, Twitter Automator delay, Twitter Automator hashtag as link prefix

+++ appseed/aura/aura/aura/aura_smart_pointer1.h
@@ -574,7 +574,7 @@ class sec_cotaskptr :
 public:
 
 
-   SIZE_T         m_size;
+   DWORD         m_size;
 
 
    sec_cotaskptr()
